feat(spring): split embedding autoconfiguration and register onnx fallback
Extract SpectorEmbeddingAutoConfiguration to register EmbeddingProvider
beans before SpectorAutoConfiguration, add defaultFallbackOnnxEmbeddingProvider
with 384 dimensions, and auto-align memory index dimensions to embedder dimensions.

Refs #775
