+++ synapse/spector-spring/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,1 +1,2 @@
+com.spectrayan.spector.spring.autoconfigure.SpectorEmbeddingAutoConfiguration
 com.spectrayan.spector.spring.autoconfigure.SpectorAutoConfiguration
